test: split httpCall rewrite coverage cases (#297)
